Add files via upload
This project is a SQL-based analysis of the COVID-19 dataset. The project aims to provide further insights from the dataset,

By the end of this project, you will have gained practical experience in using SQL for data analytics, and you will have contributed to the ongoing efforts to comprehend and combat the COVID-19 pandemic through data-driven insights.